scripts/requirements.txt
onnx>=1.14.0
onnxruntime>=1.15.0
torch>=1.13.0
torchvision>=0.14.0

Pillow>=9.0.0
numpy>=1.21.0

requests>=2.28.0

cerebrium>=1.0.0

matplotlib>=3.5.0
opencv-python>=4.7.0

